show banner if pr not merged or build not completed
